Time used for incremental build when changing a single .cc file:
- master: 26s
- timestamp check for library update: 9s
- timestamp check + gold linker: 5s

So overal build time reduces from 26s to 5s.

@keith-bennett-airmap